pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "dppeeper"
version = "0.0.8"
description = "Tool to manually analyze ICs with the dupico"
authors = [
   { name = "Fabio Battaglia", email = "hkzlabnet@gmail.com" }
]
keywords = ["dupico", "DuPAL", "analyze"]
readme = "README.md"
license = { text = "CC BY-SA 4.0 DEED" }
classifiers = [
    "Programming Language :: Python :: 3"
]
requires-python = ">=3.12"
dependencies = [
    "pyserial ~= 3.5",
    "dupicolib >= 0.4.2",
    "dpanalib >= 0.0.1"
]

[project.scripts]
dppeeper = "dppeeper.frontend:cli"

[tool.setuptools.packages.find]
where = ["src", "."]  # list of folders that contain the packages (["."] by default)
include = ["dppeeper*", "resources*"]
exclude = ["test*"]

[project.urls]
repository = "https://github.com/DuPAL-PAL-DUmper/dppeeper"